+++ Model/Command/Capture.php
@@ -4,18 +4,23 @@ namespace Unzer\PAPI\Model\Command;
 
 use Magento\Checkout\Model\Session;
 use Magento\Framework\Exception\LocalizedException;
-use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\UrlInterface;
 use Magento\Payment\Gateway\Command\ResultInterface;
 use Magento\Payment\Model\InfoInterface;
+use Magento\Sales\Api\Data\TransactionInterface;
 use Magento\Sales\Model\Order;
 use Magento\Sales\Model\Order\Payment as OrderPayment;
 use Magento\Sales\Model\Order\Payment\Transaction;
 use Magento\Sales\Model\Order\Payment\Transaction\BuilderInterface;
 use Magento\Sales\Model\Order\Payment\TransactionFactory;
 use Magento\Store\Model\StoreManagerInterface;
+use Magento\Vault\Api\Data\PaymentTokenInterface;
+use Magento\Vault\Model\Ui\VaultConfigProvider;
 use Psr\Log\LoggerInterface;
 use Unzer\PAPI\Model\Config;
+use Unzer\PAPI\Model\Method\Base;
+use Unzer\PAPI\Model\Vault\VaultDetailsHandlerManager;
+use UnzerSDK\Constants\RecurrenceTypes;
 use UnzerSDK\Exceptions\UnzerApiException;
 use UnzerSDK\Resources\AbstractUnzerResource;
 use UnzerSDK\Resources\TransactionTypes\Authorization;
@@ -40,24 +45,26 @@ use UnzerSDK\Resources\TransactionTypes\ChargeFactory;
  * limitations under the License.
  *
  * @link  https://docs.unzer.com/
- *
- * @package  unzerdev/magento2
  */
 class Capture extends AbstractCommand
 {
     /**
      * @var BuilderInterface
      */
-    private $_transactionBuilder;
+    private BuilderInterface $_transactionBuilder;
 
     /**
      * @var ChargeFactory
      */
-    private $chargeFactory;
+    private ChargeFactory $chargeFactory;
+
+    /**
+     * @var VaultDetailsHandlerManager
+     */
+    private VaultDetailsHandlerManager $vaultDetailsHandlerManager;
 
     /**
      * @inheritDoc
-     * @param TransactionFactory $_transactionFactory
      */
     public function __construct(
         Session $checkoutSession,
@@ -67,12 +74,14 @@ class Capture extends AbstractCommand
         UrlInterface $urlBuilder,
         BuilderInterface $transactionBuilder,
         StoreManagerInterface $storeManager,
-        ChargeFactory $chargeFactory
+        ChargeFactory $chargeFactory,
+        VaultDetailsHandlerManager $vaultDetailsHandlerManager
     ) {
         parent::__construct($checkoutSession, $config, $logger, $orderHelper, $urlBuilder, $storeManager);
 
         $this->_transactionBuilder = $transactionBuilder;
         $this->chargeFactory = $chargeFactory;
+        $this->vaultDetailsHandlerManager = $vaultDetailsHandlerManager;
     }
 
     /**
@@ -113,6 +122,13 @@ class Capture extends AbstractCommand
             throw new LocalizedException(__('Failed to charge payment.'));
         }
 
+        $methodInstance = $payment->getMethodInstance();
+        if ($this->isVaultSaveAllowed($methodInstance)) {
+            $paymentMethodCode = $methodInstance->getCode();
+            $this->vaultDetailsHandlerManager->getHandlerByCode($paymentMethodCode)
+                ->handle($commandSubject['payment'], $charge);
+        }
+
         $this->_setPaymentTransaction($payment, $charge);
         return null;
     }
@@ -125,7 +141,7 @@ class Capture extends AbstractCommand
      * @param float $amount
      * @param string|null $storeId
      * @return Charge
-     * @throws UnzerApiException
+     * @throws UnzerApiException|LocalizedException
      */
     protected function _chargeExisting(Order $order, string $paymentId, float $amount, string $storeId = null): Charge
     {
@@ -153,12 +169,14 @@ class Capture extends AbstractCommand
      * @param InfoInterface $payment
      * @param float $amount
      * @return Charge
-     * @throws UnzerApiException
      * @throws LocalizedException
-     * @throws NoSuchEntityException
+     * @throws UnzerApiException
      */
-    protected function _chargeNew(Order $order, InfoInterface $payment, float $amount): Charge
-    {
+    protected function _chargeNew(
+        Order $order,
+        InfoInterface $payment,
+        float $amount
+    ): Charge {
         $storeId = $order->getStoreId();
 
         $currency = $order->getBaseCurrencyCode();
@@ -175,9 +193,31 @@ class Capture extends AbstractCommand
         ]);
         $charge->setOrderId($order->getIncrementId());
 
-        return $this->_getClient($storeId, $order->getPayment()->getMethodInstance())->performCharge(
+        $unzerClient = $this->_getClient(
+            $storeId,
+            $order->getPayment()->getMethodInstance()
+        );
+
+        $isSaveToVaultActive = $payment->getAdditionalInformation(
+            VaultConfigProvider::IS_ACTIVE_CODE
+        );
+
+        /** @var PaymentTokenInterface $vaultPaymentToken */
+        $vaultPaymentToken = $payment->getExtensionAttributes()->getVaultPaymentToken();
+
+        if ($isSaveToVaultActive || $vaultPaymentToken) {
+            $charge->setRecurrenceType(RecurrenceTypes::ONE_CLICK);
+        }
+
+        if ($vaultPaymentToken) {
+            $paymentTypeId = $unzerClient->fetchPaymentType($vaultPaymentToken->getGatewayToken());
+        } else {
+            $paymentTypeId = $this->_getResourceId($payment, $order);
+        }
+
+        return $unzerClient->performCharge(
             $charge,
-            $this->_getResourceId($payment, $order),
+            $paymentTypeId,
             $this->_getCustomerId($payment, $order),
             $this->_orderHelper->createMetadataForOrder($order),
             $this->_orderHelper->createBasketForOrder($order),
@@ -194,8 +234,6 @@ class Capture extends AbstractCommand
     ): void {
         parent::_setPaymentTransaction($payment, $resource);
 
-        $parentTransactionId = null;
-
         if ($resource->getPayment()->getAuthorization()) {
             $parentTransactionId = $resource->getPayment()->getAuthorization()->getId();
         } else {
@@ -208,7 +246,7 @@ class Capture extends AbstractCommand
                 ->setTransactionId($parentTransactionId);
 
             /** @var Transaction $parentTransaction */
-            $parentTransaction = $this->_transactionBuilder->build(Transaction::TYPE_AUTH);
+            $parentTransaction = $this->_transactionBuilder->build(TransactionInterface::TYPE_AUTH);
             $parentTransaction->setIsClosed(false);
 
             // Make sure we reset the builder since it may be reused and could override data in our transaction.